Moving some stuff around and getting links working
